Created `mod rendering`, and reorganized related files into it. `mod rendering` now contains `mod image`, `mod camera`, and `mod renderer`. Flattened out the exports in the `mod.rs` file.
Created `mod materials`, and moved `materials.rs` to the module.
Flattened out the module in the corresponging `mod.rs`

Changed imports in `mod objects` to use `super::` instead of
`crate::objects::` for the internal imports.
