+++ src/main.rs
@@ -1,12 +1,11 @@
-use lumen::camera::*;
-use lumen::Material;
-use lumen::ObjectList;
-use lumen::Renderer;
-
 use std::env;
 use std::fs::File;
 use std::io::Write;
 use std::path::PathBuf;
+
+use lumen::Material;
+use lumen::ObjectList;
+use lumen::rendering::*;
 
 fn setup() -> (Camera, ObjectList, usize) {
     let ground = Material::new_diffuse(0.5, 0.5, 0.5);
