+++ src/materials/mod.rs
@@ -1,0 +1,6 @@
+// modules
+mod materials;
+
+// flatten
+pub use materials::Material;
+pub use materials::Scatter;
