+++ src/objects/mod.rs
@@ -1,14 +1,18 @@
 // modules
-pub mod cube;
-pub mod objects;
-pub mod quad;
 pub mod record;
+
+pub mod objects;
+
 pub mod sphere;
+pub mod cube;
+pub mod quad;
 
 // flatten
-pub use cube::Cube;
+pub use record::HitRecord;
+
 pub use objects::ObjectList;
 pub use objects::Physical;
-pub use quad::Quad;
-pub use record::HitRecord;
+
 pub use sphere::Sphere;
+pub use cube::Cube;
+pub use quad::Quad;
