+++ src/rendering/camera.rs
