+++ src/rendering/image.rs
