+++ src/rendering/mod.rs
@@ -1,0 +1,15 @@
+// modules
+pub mod image;
+pub mod camera;
+pub mod renderer;
+
+// flatten
+pub use image::Image;
+pub use image::Pixel;
+
+pub use camera::Camera;
+pub use camera::CameraBuilder;
+
+pub use renderer::Renderer;
+
+
